// src/modules/notifications/interfaces/notification-payload.interface.ts
export enum NotificationType {
  INFO = 'info',
  SUCCESS = 'success',
  WARNING = 'warning',
  ERROR = 'error',
  TRANSACTION = 'transaction',
  SOCIAL = 'social',
  SYSTEM = 'system',
}

export enum NotificationChannelType {
  IN_APP = 'in-app',
  EMAIL = 'email',
  SMS = 'sms',
  PUSH = 'push',
  WHATSAPP = 'whatsapp',
  SLACK = 'slack',
}